+++ lib/simple_sizer.dart
@@ -1,22 +1,33 @@
 library simple_sizer;
 
+//From Folder appcomponents
+export 'src/exports/app components/bottom_sheets.dart';
+export 'src/exports/app components/dailog_boxs.dart';
+
+//From Folder creators
+export 'src/exports/creators/text_creator.dart';
+export 'src/exports/creators/decoration_creator.dart';
+
 //From Folder enums
 export 'src/exports/enums/text_align_enum.dart';
 export 'src/exports/enums/align_cntr_enum.dart';
 
-//From Folder extensions
-export 'src/exports/extensions/build_ctx.dart';
-export 'src/exports/extensions/align_cntr_ext.dart';
-export 'src/exports/extensions/expanded_and_flexible_ext.dart';
-export 'src/exports/extensions/fractionally_box.dart';
-export 'src/exports/extensions/padding_ext.dart';
-export 'src/exports/extensions/aspectratio_ext.dart';
-export 'src/exports/extensions/sizedbox_ext.dart';
-export 'src/exports/extensions/fitted_box_ext.dart';
-export 'src/exports/extensions/limitedbox_ext.dart';
-export 'src/exports/extensions/stba_ext.dart';
-export 'src/exports/extensions/text_ext.dart';
-export 'src/exports/extensions/boxdecor_ext.dart';
+//From Folder flutter-widgets
+export 'src/exports/extensions/flutter-widgets/build_ctx.dart';
+export 'src/exports/extensions/flutter-widgets/align_cntr_ext.dart';
+export 'src/exports/extensions/flutter-widgets/expanded_and_flexible_ext.dart';
+export 'src/exports/extensions/flutter-widgets/fractionally_box.dart';
+export 'src/exports/extensions/flutter-widgets/padding_ext.dart';
+export 'src/exports/extensions/flutter-widgets/aspectratio_ext.dart';
+export 'src/exports/extensions/flutter-widgets/sizedbox_ext.dart';
+export 'src/exports/extensions/flutter-widgets/fitted_box_ext.dart';
+export 'src/exports/extensions/flutter-widgets/limitedbox_ext.dart';
+export 'src/exports/extensions/flutter-widgets/stba_ext.dart';
+export 'src/exports/extensions/flutter-widgets/text_ext.dart';
+
+//From Folder data-types
+export 'src/exports/extensions/data-types/bool_ext.dart';
+export 'src/exports/extensions/data-types/string_ext.dart';
 
 //From Folder helper_methods
 export 'src/exports/helper_methods/shape_helper.dart';
@@ -28,9 +39,5 @@ export 'src/exports/reactive_widgets/reactive_destination.dart';
 export 'src/exports/reactive_widgets/reactive_row_col.dart';
 export 'src/exports/reactive_widgets/reactive_home.dart';
 
-//From Folder widgets
-export 'src/exports/widgets/bottom_sheets.dart';
-export 'src/exports/widgets/dailog_boxs.dart';
-
-//From Folder exports
-export 'src/exports/size_config.dart';
+//From Folder responsiveness
+export 'src/exports/responsiveness/size_config.dart';