+++ lib/src/exports/extensions/data-types/bool_ext.dart
@@ -1,0 +1,9 @@
+extension BoolExt on bool {
+  /// The `bool get isFalse => (this == false);` is a getter method defined in the `BoolExt` extension
+  /// on the `bool` type in Dart.
+  bool get isFalse => (this == false);
+
+  /// The `bool get isTrue => (this == true);` is a getter method defined in the `BoolExt` extension on
+  /// the `bool` type in Dart.
+  bool get isTrue => (this == true);
+}