+++ lib/src/exports/extensions/flutter-widgets/text_ext.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/widgets.dart';
+import 'package:simple_sizer/src/exports/creators/text_creator.dart';
+
+extension TextExt on Text {
+  TextCreator get txt => TextCreator(data: data!, style: style);
+}