+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 1.0.0
+
+- [`Breaking Change`]
+
+  The Text extensions have been `deprecated` and replaced by all new `TextCreator` class
+  for better control over ui rendering.
+
+- [`Breaking Change`]
+
+  The BoxDecoration extensions have been `deprecated` and replaced by all new `Decorate` class
+  for better control over ui rendering.
+
+- New extention added on `bool` and `String`
+
 ## 0.9.1
 
 - readme.md updated